scripts/test/requirements.txt
boto3==1.37.15
folium==0.19.5
pandas==2.0.3
python-dotenv==1.1.0
Requests==2.32.3
Shapely==2.1.0
